Split search cpp into two
